+++ src/training_set_utils.h
@@ -5,4 +5,4 @@ struct Training_Input {
 	double measured_value = 0;
 };
 
-bool create_training_set(const std::string& db_name, size_t input_layer_count, double minutes_prediction, std::vector<Training_Input>& training_set);
+bool create_training_set(const std::string& db_name, size_t input_layer_count, size_t minutes_prediction, std::vector<Training_Input>& training_set);
